fix #59

+++ timflow/steady/well.py
@@ -909,7 +909,8 @@ class WellStringBase(Element):
         # include resistance by finding position of element in coefficient matrix
         # and subtracting resfac (resistance factor).
         iself = self.model.elementlist.index(self)
-        jcol = np.sum([e.nunknowns for e in self.model.elementlist[:iself]])
+        # needs to be converted to int as sum of empty list returns real
+        jcol = np.sum([e.nunknowns for e in self.model.elementlist[:iself]]).astype(int)
         irow = 0
         for w in self.wlist:
             mat[irow : irow + w.nlayers, jcol : jcol + w.nunknowns] -= w.resfac[
